display GAQ summary
