+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewModel.js
@@ -34,6 +34,8 @@ export class RunsPerDataPassOverviewModel extends RunsOverviewModel {
         this._dataPass.bubbleTo(this);
         this._qcSummary$ = new ObservableData(RemoteData.NotAsked());
         this._qcSummary$.bubbleTo(this);
+        this._gaqSummary$ = new ObservableData(RemoteData.NotAsked());
+        this._gaqSummary$.bubbleTo(this);
 
         this.patchDisplayOptions({ horizontalScrollEnabled: true });
     }
@@ -66,6 +68,20 @@ export class RunsPerDataPassOverviewModel extends RunsOverviewModel {
         }
     }
 
+    /**
+     * Fetch GAQ summary for given data pass
+     * @return {Promise<void>} resolves once data are fetched
+     */
+    async _fetchGaqSummary() {
+        this._gaqSummary$.setCurrent(RemoteData.loading());
+        try {
+            const { data: gaqSummary } = await getRemoteData(buildUrl('/api/qcFlags/summary/gaq', { dataPassId: this._dataPassId }));
+            this._gaqSummary$.setCurrent(RemoteData.success(gaqSummary));
+        } catch (error) {
+            this._gaqSummary$.setCurrent(RemoteData.failure(error));
+        }
+    }
+
     // eslint-disable-next-line valid-jsdoc
     /**
      * @inheritdoc
@@ -76,6 +92,7 @@ export class RunsPerDataPassOverviewModel extends RunsOverviewModel {
         }
         this._fetchDataPass();
         this._fetchQcSummary();
+        this._fetchGaqSummary();
         super.load();
     }
 
@@ -119,6 +136,14 @@ export class RunsPerDataPassOverviewModel extends RunsOverviewModel {
      */
     get qcSummary() {
         return this._qcSummary$.getCurrent();
+    }
+
+    /**
+     * GAQ summary getter
+     * @return {RemoteData<GaqSummary>} GAQ summary
+     */
+    get gaqSummary() {
+        return this._gaqSummary$.getCurrent();
     }
 
     /**
