+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewPage.js
@@ -28,6 +28,7 @@ import runNumberFilter from '../../../components/Filters/RunsFilter/runNumber.js
 import { qcSummaryLegendTooltip } from '../../../components/qcFlags/qcSummaryLegendTooltip.js';
 import { isRunNotSubjectToQc } from '../../../components/qcFlags/isRunSubjectToQc.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
+import { getQcSummaryDisplay } from '../ActiveColumns/getQcSummaryDisplay.js';
 
 const TABLEROW_HEIGHT = 59;
 // Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
@@ -62,6 +63,7 @@ export const RunsPerDataPassOverviewPage = ({
         detectors: remoteDetectors,
         dataPass: remoteDataPass,
         qcSummary: remoteQcSummary,
+        gaqSummary: remoteGaqSummary,
         displayOptions,
         dataPassId,
     } = perDataPassOverviewModel;
@@ -80,12 +82,28 @@ export const RunsPerDataPassOverviewPage = ({
                 : {},
             Other: () => {},
         }),
-        globalAggregatedQuality: {
-            name: 'GAQ',
-            visible: true,
-            format: (_, { runNumber }) => frontLink(h('button.btn.btn-primary.w-100', 'GAQ'), 'gaq-flags', { dataPassId, runNumber }),
-            profiles: ['runsPerDataPass'],
-        },
+        ...remoteGaqSummary.match({
+            Success: (gaqSummary) => ({
+                globalAggregatedQuality: {
+                    name: 'GAQ',
+                    visible: true,
+                    format: (_, { runNumber }) => {
+                        const runGaqSummary = gaqSummary[runNumber];
+                        const gaqDisplay = runGaqSummary
+                            ? getQcSummaryDisplay(runGaqSummary)
+                            : h('button.btn.btn-primary.w-100', 'GAQ');
+
+                        return frontLink(
+                            gaqDisplay,
+                            'gaq-flags',
+                            { dataPassId, runNumber },
+                        );
+                    },
+                    profiles: ['runsPerDataPass'],
+                },
+            }),
+            Other: () => ({}),
+        }),
         ...createRunDetectorsAsyncQcActiveColumns(
             remoteDetectors.match({
                 Success: (payload) => payload,
